--- src/components/index.js
@@ -1,2 +1,0 @@
-const Component = () => <h1>Hello</h1>
-export default Component
